Extract shared camelToKebab and write utilities
Move duplicated camelToKebab from component-css.js and component-js.js
into src/util/camel-to-kebab.js. Move duplicated write helper from
init.js and create.js into src/util/write.js.
